Damage/Exhaust/Attacked update /w bug fixes
Implemented code that correctly recognizes when Damage Taken, the
Exhaust flag, or the numAttacked flag is set.  Now we just need to parse
the line and get the information in the correct place.

Added debug lines to indicate when backtracing begins and ends.

Corrected a bug that was causing the parser to read incomplete lines
from the log file.