Parser now handles Damage/Exhaust/Attacked
Added functionality to parser and helpers that parses out the
information useful for updating stat changes.

Created a new Proto Class, update.proto.  This class consists of an
instruction and a list of arguments to be handled by the Handler so we
don't have to pipe around the full board.  Also implemented the proto
class, so now it will flood all of your consoles with "useful" Debugging
information.  Enjoy.

TODO : Continue implementing the proto class.  Will be moving most of
the logic to the handler.